My first (presumably) working implementation of the third task. Haven't tested it completely through.
It's the fastest solution I came up with. There is probably a better and "more elegant" way to solve the third task.

Have to get rid of some of the trash and out of place comments later on. Considering reinstating the BBL subproject, since I am using so called "helper models" -- optimized domain entity versions for certain tasks.

Estimated time in total? No idea.. 6 to 12 hours?